+++ src/components/atoms/TextEditor/index.tsx
@@ -1,0 +1,21 @@
+import React, { useState } from "react";
+import { Editor } from "react-draft-wysiwyg";
+import { EditorState } from "draft-js";
+import "react-draft-wysiwyg/dist/react-draft-wysiwyg.css";
+
+const TextEditor: React.FC = () => {
+  const [editorState, setEditorState] = useState(EditorState.createEmpty());
+
+  const onEditorStateChange = (newEditorState: EditorState) => {
+    setEditorState(newEditorState);
+  };
+
+  return (
+    <Editor
+      editorState={editorState}
+      onEditorStateChange={onEditorStateChange}
+    />
+  );
+};
+
+export default TextEditor;